requirements.txt
Pillow==10.0.0
keyboard==0.13.5
pystray==0.19.4
pyinstaller==6.3.0
